cleanup startserver.sh, make it executable
startserver.sh's execute bit was unset so trying to run the script would
yield a "permission denied" error. Added a shebang at the top of the
file so it will always run under the right shell. Also cleaned up a bit
of the code including quoting some variables to avoid some globbing
weirdness.